Added blocking & redirection functionality.
When going to the hardcoded url, popup blocks video. Other videos will
not be blocked.

Added a few todos.

Pages with videos in them will allow user to report the video in that
page by clicking on the context menu. User will be alerted if no video
exists on the page.